+++ legacy/gender-routing-response.php
@@ -3,7 +3,7 @@ require_once '_includes/functions.php';
 header("content-type: text/xml");
 echo "<?xml version=\"1.0\" encoding=\"UTF-8\"?>\n";
 
-$gender = getIvrResponse("gender-routing.php", null, [1, 2]);
+$gender = getIvrResponse("gender-routing.php", null, [VolunteerGender::MALE, VolunteerGender::FEMALE, VolunteerGender::NO_PREFERENCE]);
 $_SESSION['Gender'] = $gender;
 ?>
 <Response>