+++ legacy/gender-routing.php
@@ -7,7 +7,10 @@ echo "<?xml version=\"1.0\" encoding=\"UTF-8\"?>\n";
 <Response>
     <Gather language="<?php echo setting('gather_language') ?>" hints="<?php echo setting('gather_hints') ?>" input="<?php echo getInputType() ?>" timeout="10" speechTimeout="auto" action="gender-routing-response.php?SearchType=<?php echo $_REQUEST['SearchType']?>" method="GET">
         <Say voice="<?php echo voice(); ?>" language="<?php echo setting('language') ?>">
-            <?php echo sprintf("%s %s %s, %s, %s, %s", word('press'), word('one'), word('to speak to a man'), word('press'), word('two'), word('to speak to a woman'))?>
+            <?php
+                $gender_no_preference = (setting("gender_no_preference") ? sprintf(", %s %s %s", getPressWord(), word('three'), word('speak no preference')) : "");
+                echo sprintf("%s %s %s, %s %s %s%s", getPressWord(), word('one'), word('to speak to a man'), getPressWord(), word('two'), word('to speak to a woman'), $gender_no_preference)
+            ?>
         </Say>
     </Gather>
 </Response>