+++ RELEASENOTES.md
@@ -2,6 +2,7 @@
 
 ### 4.0.0 (UNRELEASED)
 * Voicemail deletion feature. [#528]
+* Gender routing with no preference option. [#523]
 * Fix for unspecified language volunteers not correctly using fallback language setting.  The new fallback is whatever the first setting is in the `language_selections` setting.  [#507]
 * Fix for music on hold not inheriting from redirected service body volunteer routing. [#532]
 * Fix for groups page dropdown when no service body is selected. [#505]